Split Reddit objects into "things"
This follows the parlance of the Reddit API and the old source code that
used to be publicly available. I'm not sure if all this indirection is
really necessary -- it's questionable what the role of Redditor is now
-- but this is how I structured things in the old Ruby version of
usaidwat, and that worked pretty well, so I'll use that as a starting
point and iterate over time.