Fix ownership issues with Vec<Comment> and Vec<Submission>
While writing tests for Redditor, I noticed that I could not write

    let posts = test_client().submissions();

because of an issue with borrowing the return value of `submissions()`
and `comments()`. Instead, I had to write this

    let client = test_client();
    let posts = client.submissions();

which is not ergonomic, to say the least.

Turns out I will need to clone the vectors of comments and submissions
(I could use Rc or Arc, but that's a little unwieldy for the purposes of
this program).

Thanks to a suggestion from ChatGPT, I also modified the methods to
return Iterators instead of Vecs, which should further make them a bit
easier to use.