Profile created_utc is now a floating-point number
When I wrote the original Ruby code 13 years ago, the created_utc
timestamp returned by /users/<username>/about.json was an integer. Now
it's a floating-point number, so we have to parse it as such, or
serde_json will (quite reasonably) barf.

This also means that we no longer have to handle the deserialization of
integers to DateTimes at all, so we can remove that code.

This does unfortunately mean that usaidwat will likely break as the data
returned by the Reddit API changes, but we'll cross that bridge when we
come to it.