Reddit vote tallies and scores are signed integers
It makes sense that the overall score is a signed integer and listing it
as a u64 was a mistake. However, I'm surprised to find that the "ups"
and "downs" tallies are also signed. I don't know what a negative "ups"
means, since you'd think that given a score of, e.g., -3, the "ups"
would be 0 and the "downs" would be -3, or whatever, but for some
reason, Reddit uses signed integers.

Kind of a shame that I have to update the comment and post listings that
have been used since 2012 -- they were a sort of time machine into my
earlier days on Reddit -- but it's good to test against the latest API
formats.