Provide options to Viewable::view()
We'll need these options when printing comment logs, since comment logs
can be printed in "raw" or "oneline" mode.

I don't especially love this pattern because it will get cumbersome to
add additional options in the future. I feel like there is a better
pattern to use here...Decorator or Builder or something? But this will
be fine for now.