Group comments by day and hour using Timeline struct
This can be used to print a matrix showing when a user has commented,
grouped by day and hour -- basically, a rough summary of their activity,
much like GitHub's activity viewer.

I virtually never actually use this feature, but it is a part of the
current Ruby version of usaidwat, and I guess it's kind of cool that it
works, so I'll keep it.