Print "user has no comments" instead of empty timeline
This mimics the behavior of the current Ruby version of usaidwat.

It's hard to test, though. There's probably a way to test all the CLI
logic, but I should really move this conditional check into a testable
function or method to ensure the proper output. I could put it into
Timeline::view(), but I don't think it's really appropriate there, so
I'll leave the testability of this logic as a future exercise.