Parameterize Redditor by time source
For tests, we will have to mock time so that we can always test against
a known time and time strings. In Ruby, we could use some monkey-patching
shenanigans to mock system time in tests, but this is not possible, or
at least not kosher, in Rust. Instead, we will pass a Clock to Redditor
which it can use to get the time. It's a bit cumbersome to parameterize
Redditor in this fashion, but it solves the problem nicely.

We could also have Redditor take a Box<dyn Clock> instead of a generic
parameter, but the generic gives us an additional bit of type safety
with little cost.