Refactor Redditor::age() to take a clock
Time is only used to calculate age, so there's not really a need to
store the clock as an instance variable (and parameterize Redditor in
the process). It's slightly annoying to have to pass in a clock every
time age() is called, but hey, that's probably how we'd do it in my
beloved Haskell.

Furthermore, I am going to have to calculate the age of other entities
in the system, such as Comments, and I don't really want to have to
parameterize _every_ entity with an "age" and store an instance
variable, so passing it into the method where it is used will be less
cumbersome in the long run.