Implement tally and posts tally commands
The code around this is a little wonky: namely, it's cumbersome to
implement view() on an iterator (specifically, IntoIter) because we have
to consume the iterator to iterate through it, and view() does not
borrow its self parameter; also, we have to use the iterator _twice_,
once to find the longest subreddit (for formatting purposes) and once to
actually iterate through the items to format them. The code works fine
so it will suffice for now, but it is far from elegant.