Horizontally scroll oneline posts
This mimics the behavior of the Ruby tool, which neither chops nor wraps
long lines, but rather forces them to be scrolled horizontally in the
pager. There are a bunch of hoops to go through with environment
variables to make this work, and I'm not really sure I am doing the
right thing here. The Ruby tool uses the value of $LESS, defaulting to
$FSRX, unless --oneline is specified, in which case $RS is hardcoded.
That presents a nice default interface but doesn't fully respect the
user's preferences...but I also practically require -R to be set or else
the user will see ugly ANSI escape codes. This logic will work for now
but I might want to revise it in the future, although it is more or less
what I have been doing in the Ruby tool for 13 years without complaint
(of course, I am probably the only user of the Ruby tool).