Colored text is only output for ttys
Turns out this was already working, but I have $CLICOLOR_FORCE=true set,
and the colored crate reasonably respects that. There is actually a
whole hierarchy of environment variables that are set, and in fact it is
a little wonky, because I had to set CLICOLOR_FORCE=0 to test this --
not CLICOLOR_FORCE=false. But it does appear to work reasonably as
expected, so I should not have to handle this myself.

Unfortunately, _testing_ this functionality is a bit hard. I suspect
that colored::control::set_override is not threadsafe, and Rust test
cases are probably run in parallel, so trying to force an override
breaks tests that _are_ testing for color. I suppose I should set an
override = true for _every_ test for colorized output, but that is a bit
annoying, so I'll just skip testing for non-colorized output for now.