Use 'a instead of 'static lifetime
'static works, but it's not what we really want per se. Probably what we
really want is to take ownership of the suffix, but restricting the
lifetime of TestService to live at least as long as its suffix also
works fine...partly because its suffix is in fact a static string. But
whatever.