Mark Rust v1.88 as the lowest supported version
Pretty crazy that it just came out 3 days ago, because I am in fact
using a piece of syntax that does not compile with Rust v1.87: a `if
let` alongside an `if` statement.

    if !opts.raw && let Some(grep) = &opts.grep

This is "unstable" in previous versions of Rust and will not compile, so
we _must_ use Rust v1.88.

  [1]: https://github.com/rust-lang/rust/pull/132833