Fix relativetime bug
We use the relativetime crate to calculate human-readable strings for
ages. Unfortunately, it sometimes returns things like "1 months ago".
Even more unfortunately, that crate is no longer developed -- it doesn't
even exist on GitHub anymore -- so we're not likely to see any updates
to it anytime soon. For now, let's just hack around the bug, until we
can come up with a better solution.