Implement a bare Markdown parser
We want to be able to take a Redditor's comments and pass them to an LLM
for summarization and sentiment analysis. To do so, we have implemented
a simple parser that strips out nearly all markup and other text
decorations, separating the content of a blob of Markdown from its
markup, so we can pass the raw comments of Markdown text to an LLM.

Truth be told, an LLM could probably deal with Markdown markup just fine
-- we could probably prompt it by telling it to ignore any Markdown
markup -- but this ensures we are getting a "clean" blob of text for
processing.