Auth methods should take Into<String>
Into<String> is more flexible, as it allows for anything that can be
converted into a String, including both String and &str.

I should consider changing other methods in the codebase that take
String or &str to take Into<String> instead, but it turns out that is
going to be a massive job.