Actually, unnest imports
I think it can make sense to nest imports when importing a bunch of
modules from a single crate, like

    use std::{env, error, fmt}

or

    use crate::{markdown, text}

And definitely when importing a list of functions or structs from a
single crate.

But in general, I find that more complex nesting is hard to read. It
also makes it more annoying to "pub use" items, and RustRover doesn't
like the style, so prefer the unnested form unless the nested form is
short and clear.