Make HTTP calls asynchronously
Asynchronous calls are the expected, default pattern when using reqwest,
and it should improve the performance of the program as it will be able
to make all three necessary HTTP calls in parallel once they are being
made asynchronously.

Unfortunately, at the moment there is a panic at runtime, but overall,
the code works -- and tests pass.