Implement a tokio-friendly pager
Tokio emitted an error about unclosed or invalid file descriptors when
run in conjunction with a command that sent its output to a pager, such
as `log` or `posts log`. This was due to the fact that the pager crate
opened up a new thread to run the pager, and may have been exacerbated
by the fact that the pager crate actually ran the pager in the parent
process, not the child. Regardless, it was necessary to implement a
pager in such a way that it would play nicely with tokio.

This new pager is spawned in a process controlled by tokio, and calling
code properly awaits on it, fixing the file descriptor bug emitted
before by tokio.