Record available OpenAI models and usage costs
I may not end up supporting all of these models -- I may even end up
picking the best one on my own and restricting the OpenAI API client to
use that one -- but for development, it will be nice to be able to
switch between them.

It will also be nice to know what the cost of each model is, and perhaps
in the future, I can make that option discoverable and configurable by
the user. It might also end up being unnecessary and I might ultimately
strip it out, but we'll see what happens as the ai module evolves.