Create OpenAI API client with HTTP client implementation
This will be necessary when we write tests, as we will have to pass in a
mocked HTTP client to implement deterministic tests. We don't yet have a
general type for the client so we'll use reqwest::Client as the type,
but this will definitely change in the future.

"client" would be a better name for this parameter, but for parity with
the Redditor data structure, we'll refer to it as an HTTP "service".