Refactor HTTP client into separate module
We are going to need this for other HTTP services, such as communicating
with AI APIs, so it makes sense to create a top-level HTTPClient to
define basic and general HTTP operations, and then implement lower-level
clients as needed.

We might end up refactoring this evening more, perhaps by creating a
default implementation of get(), but this more limited refactoring will
set us on the path to success.