Move HTTPService::get() back to RedditService
I had defined it on the "base" HTTPService because it would be useful
for the OpenAI API client I am writing...but that client communicates
with the OpenAI API via HTTP POST, so a get() method isn't actually
necessary to implement for that trait. Maybe someday it will be, but for
now, it makes sense to move it back to RedditService.