Use the cheapest model to generate a summary
I had already spent 50¢ to test usaidwat today! Okay, that's not a lot,
but I guess it could add up, so let's try using the cheapest (and
fastest) model to generate summaries instead.

Ultimately we should allow users to specify preferences using flags like
`--fast` or `--best`, or maybe directly specify a model with `--model`,
but let's see how good the summaries are with the cheapest, fastest
model.