"Prevent" AI from identify text as social media posts
The AI often prefaced its response with lines like "In this series of
social media posts..." But obviously anyone using this tool already
knows these are social media posts, so try to get the AI to refrain from
doing that.

This has only been a bit successful, and was not as much of a problem
when using the default model (Model::default()), and I'm not sure the
output with this model and prompt is optimal, so continue iterating on
this.