Revert "Test summary using an actual OpenAIResponse"
Actually...we won't have to do this. Concatenation of a response will
occur on the OpenAIResponse struct itself, so we can test there to
ensure that the output is correctly concatenated, and can continue to
use our own custom response.

We'll want to use our own custom response because we'll probably have to
spy on it.

This reverts commit 8ae5c5e2ce29f60882cdaca6565894c4091728e5.