License under the Apache License 2.0
I always agonize over the licensing of my projects, perhaps more than
the code itself. Generally I like to go with the GPLv3 for _programs_
(as opposed to libraries) because I think offers the most freedom to end
users to continue to use and distribute my beautiful program forever;
however, there is some substantial, not to mention useful, library code
in usaidwat that could be extracted out into separate crates, and I like
the idea that others may make use of this code in their own programs.
Furthermore, the Apache License 2.0 + MIT is more common in the Rust
world; I don't see the need to dual-license, but the Apache License 2.0
seems to be good enough, offering the most freedom with the most
flexibility and compatibility with other Rust libraries.

I was going to add an exception that this code cannot be used as
training material for an LLM, but I have no faith whatsoever that Sam
Alterman and his ilk give a shit about licensing and copyright, so such
an exception wouldn't be worth the bytes it was written in, nor the time
it would take me to type it.