Release v2.0.0
New usaidwat, who dis?

This completes a two-month effort to rewrite my old usaidwat [1] tool in
Rust, after growing frustrated by maintaining a Ruby tool in a dying
ecosystem. I wanted to be able to compile and deploy a single static
binary that I can easily copy from one machine to another, and Rust
allowed me to do that in a matter of months. Thank you, Rust!

The repository still isn't public, and I don't have a homepage (and
probably never will), but this is good enough for now.

[1]: https://github.com/mdippery/usaidwat