Update repository URL
Yes, I am now hosting this repository on GitHub. The first new
repository I have created on GitHub since 2022. While I was an early
champion of GitHub, I have been reluctant to continue using it, as it
has just become a repository of fuel for LLMs. But what other choice do
I have? It's the best way for usaidwat's users to get usaidwat's source
code, and it's the easiest way for developers to contribute. GitHub
Issues is nice, even if Microsoft continues to enshittify the product
with enterprise slop.

Microsoft and Sam Altman and all of them may not respect copyright nor
licensing enough to refrain from stealing my code, but that's going to
happen regardless of whether I put it on GitHub, or elsewhere online,
and no matter what license I post it under -- even one that forbids its
use in training LLMs. And with everyone pushing entire corporate
monorepos up to Cursor and Copilot and whatever else, my personal
decision to hoard my code doesn't mean a thing in a practical sense. So
here it is, world: take it if you will.