Remove new line characters from Reddit post titles
For some reason, Reddit sometimes returns comment post titles with an
embedded new line. In this test case, for example, the comment post
title is "[OC] \n...", which made the output of `usaidwat log` a little
wonky: namely, it was not colored properly. This is an actual test
case from Reddit.

The solution is simple: we will now just remove new line characters from
output.