Make an elided lifetime explicit
Here, self has an implicit lifetime of 'a, but that lifetime was elided
from the return value. That issue is now flagged in the latest Rust
compiler. It's pretty simple to fix: we need an explicit lifetime for
the returned value. We can use '_ which means "the same lifetime as the
self parameter".