Handle GPT-5 responses
GPT-5 adds an output type called "reasoning" with a "summary" field. We
don't use it for anything right now, but we still have to handle it, or
else serde_json can't deserialize the OpenAI response into a struct.
Thus, OpenAIOutput is now an enum with Reasoning and Message elements.

Return a string in OpenAIOutput::content() might not really make sense
in the context of a Reasoning response, and it might be best to return
an Option<String> instead, but we'll continue returning a String for
now.