Default to GPT-4.1 nano
GPT-5 nano is much, much slower than GPT-4.1 nano, and to be honest, I
don't like the output anyway; I think the GPT 4.1 output (across all
variants) is better. But GPT-5 is cheaper, and it is the flagship model,
so I'll use GPT-5 for the "best" and "cheapest" options, but continue to
use 4.1 nano for the "fastest".