Display available summarization models
When -vv or higher is passed to the summary command, the mapping of
models to --model options is printed. This is mostly useful as a
diagnostic tool, but it is a good way to refresh my memory of what
models are in use.

It might be nicer to _just_ show the available models and exit, but
that's a feature for later.