Test GPT-5 response output
GPT-5 responses include a reasoning summary block that is not present in
GPT-4 responses. We now handle this using an enum, but we should test
to make sure the code works.

Note that a lot of these GPT-5 responses are inferred from brief testing
and may not be entirely accurate. They were not all generated from an
actual API response using a test script.