Replace full license header with SPDX metadata
While the Apache Foundation itself stipulates that its software contains
the full Apache 2.0 license header, it is honestly a big long and ugly.
For what it's worth, the Linux kernel has moved to using
SPDX-License-Identifier metadata, and if it's good enough for the Linux
kernel, it's good enough for me -- with the caveat that I am including a
short copyright notice as well, just to make it easier to track the
providence of this file.

I am also including a barebones COPYRIGHT file to make it easy to
quickly find out the copyright information for this project. This might
eventually morph into an Apache 2.0 LICENSE file [1] but, ah, I'll cross
that bridge when I come to it.

[1]: https://www.apache.org/licenses/NOTICE-2.0.txt