Internal pager methods should be private
When pager was updated to detect whether stdout was cat or not a tty,
public visibility was mistakenly not removed from the internal methods.
Only the single `page()` method, which dispatches based on the nature of
stdout, should be exposed to callers.

There is also an issue with `SubredditCounter::from_iter()` which I will
try to address later.