Show available AI models with --help
It's helpful to easily list the models that are in use, but augmenting
--help dynamically is difficult, so we'll define the models as an "after
help" string. This means we have to hardcode the models in a string,
which is _bad_; eventually I'd like to dynamically create this list at
compilation time using a macro, but for now, this solution will suffice.