Dynamically generate available models for help string
Hot dog! It turns out that "after_help" takes an _expression_, not just
a constant string, so I can call a function (presumably as long as it
returns a String or a String-like object). That means that I _can_
easily generate the list of available models at runtime.

I would still like to not have to specify that the AI provider is
OpenAI, but this is a great change that will work for the time being.