Use new hypertyper crate
I split off usaidwat.http into a separate crate called hypertyper. I'm
not entirely convinced that this crate is really that useful, but
usaidwat.ai uses it as well as usaidwat.reddit, and usaidwat.ai will
soon be split off into a separate crate called cogito, and creating a
common crate for both usaidwat and cogito is preferable to copying
usaidwat.http from one project to another.

I might have to think about the design of usaidwat.http and its usage
more, but this a good solution for now.