Write re-exported structs, traits, and functions inline
The #[doc(inline)] directive rewrites documentation for structs, traits,
and functions in the Structs and Traits section instead of in a
"re-exported" section.

This might be a bit confusing in practice, especially since
usaidwat::markdown re-exports some functions from
usaidwat::markdown::bare, but we'll go with it for now because I really
hate the ugly (and potentially even more confusing) "re-exported" text.