Experiment with Claude instead of OpenAI
Claude is a little bit faster than OpenAI...but mostly I just want to
see if it yields better summaries. usaidwat will now make requests to
Claude instead of OpenAI.

We could probably improve the design of usaidwat to make it easier to
switch providers in the future -- there is a lot of manual work that
needs to be done at the moment -- but that can be an improvement for
later. I don't expect to switch providers often anyway.

Closes #8.