Schedule Claude changes for a v5.0 release
Originally I had planned to change the provider to Claude in v4.1, and
release some changes related to the display of self posts in v4.2, but I
think these Claude changes need more experimentation, and I might not
end up releasing them at all, so I'm going to flip the order and release
the self-post changes in v4.1.

Switching to Claude also seems like it warrants a major version bump, so
I'm going to release the Claude changes as v5.0.