Show or hide self posts in posts log subcommand
Recently a feature was added to only show self posts, but it seems
equally useful to hide all self posts, and of course it is useful to
continue with the previous behavior of showing both kinds of posts, so
--self has been updated to handle three options.

I would like --self to imply --self=only. That should be possible by
setting default_missing_value in the Clap arg...but doing so still
results in an error if a value for --self is not defined. Not a big deal
for release but something I should fix later.

Closes #12.